+++ src/main/java/roomescape/controller/MemberController.java
@@ -10,25 +10,25 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.member.MemberResponse;
 import roomescape.dto.signup.SignupRequest;
-import roomescape.service.member.MemberServiceImpl;
+import roomescape.service.member.MemberService;
 
 @RestController
 @RequestMapping("members")
 public class MemberController {
 
-    private final MemberServiceImpl memberServiceImpl;
+    private final MemberService memberService;
 
-    public MemberController(MemberServiceImpl memberServiceImpl) {
-        this.memberServiceImpl = memberServiceImpl;
+    public MemberController(MemberService memberService) {
+        this.memberService = memberService;
     }
 
     @GetMapping
     public ResponseEntity<List<MemberResponse>> getAllMember() {
-        return ResponseEntity.ok().body(memberServiceImpl.findAllMembers());
+        return ResponseEntity.ok().body(memberService.findAllMembers());
     }
 
     @PostMapping
-    public ResponseEntity<MemberResponse> addMember(@RequestBody SignupRequest signupRequest){
-        return ResponseEntity.created(URI.create("/members")).body(memberServiceImpl.addMember(signupRequest));
+    public ResponseEntity<MemberResponse> addMember(@RequestBody SignupRequest signupRequest) {
+        return ResponseEntity.created(URI.create("/members")).body(memberService.addMember(signupRequest));
     }
 }
