+++ src/main/java/roomescape/service/reservation/ReservationService.java
@@ -1,24 +1,108 @@
 package roomescape.service.reservation;
 
+import jakarta.validation.Valid;
+import java.time.LocalDateTime;
 import java.util.List;
+import org.springframework.stereotype.Service;
 import roomescape.domain.Member;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
+import roomescape.domain.Theme;
 import roomescape.dto.admin.AdminReservationRequest;
 import roomescape.dto.reservation.MemberReservationResponse;
 import roomescape.dto.reservation.ReservationRequest;
 import roomescape.dto.reservation.ReservationResponse;
 import roomescape.dto.search.SearchConditionsRequest;
+import roomescape.exception.member.MemberNotFoundException;
+import roomescape.exception.reservation.ReservationAlreadyExistsException;
+import roomescape.exception.reservation.ReservationInPastException;
+import roomescape.exception.reservationtime.ReservationTimeNotFoundException;
+import roomescape.exception.theme.ThemeNotFoundException;
+import roomescape.repository.member.MemberRepository;
+import roomescape.repository.reservation.ReservationRepository;
+import roomescape.repository.reservationtime.ReservationTimeRepository;
+import roomescape.repository.theme.ThemeRepository;
 
-public interface ReservationService {
+@Service
+public class ReservationService {
 
-    ReservationResponse create(ReservationRequest request, Member member);
+    private final ReservationRepository reservationRepository;
+    private final ReservationTimeRepository timeRepository;
+    private final ThemeRepository themeRepository;
+    private final MemberRepository memberRepository;
 
-    List<ReservationResponse> getAll();
+    public ReservationService(ReservationRepository reservationRepository,
+                              ReservationTimeRepository timeRepository, ThemeRepository themeRepository,
+                              MemberRepository memberRepository) {
+        this.reservationRepository = reservationRepository;
+        this.timeRepository = timeRepository;
+        this.themeRepository = themeRepository;
+        this.memberRepository = memberRepository;
+    }
 
-    void deleteById(Long id);
+    public ReservationResponse create(ReservationRequest request, Member member) {
+        ReservationTime reservationTime = timeRepository.findById(request.timeId())
+                .orElseThrow(() -> new ReservationTimeNotFoundException(request.timeId()));
 
-    ReservationResponse createByAdmin(AdminReservationRequest adminReservationRequest);
+        Theme theme = themeRepository.findById(request.themeId())
+                .orElseThrow(() -> new ThemeNotFoundException(request.themeId()));
 
-    List<ReservationResponse> getReservationsByConditions(SearchConditionsRequest searchConditionsRequest);
+        if (LocalDateTime.now().isAfter(LocalDateTime.of(request.date(), reservationTime.getStartAt()))) {
+            throw new ReservationInPastException();
+        }
 
-    List<MemberReservationResponse> getReservationByMember(Member member);
+        if (reservationRepository.existsByDateAndTime(request.date(), reservationTime)) {
+            throw new ReservationAlreadyExistsException();
+        }
+
+        Reservation newReservation = new Reservation(request.date(),
+                reservationTime, theme, member);
+
+        return ReservationResponse.from(reservationRepository.save(newReservation));
+    }
+
+    public List<ReservationResponse> getAll() {
+        return ReservationResponse.from(reservationRepository.findAll());
+    }
+
+    public void deleteById(Long id) {
+        reservationRepository.deleteById(id);
+    }
+
+    public ReservationResponse createByAdmin(AdminReservationRequest adminReservationRequest) {
+        ReservationTime reservationTime = timeRepository.findById(adminReservationRequest.timeId())
+                .orElseThrow(() -> new ReservationTimeNotFoundException(adminReservationRequest.timeId()));
+
+        Theme theme = themeRepository.findById(adminReservationRequest.themeId())
+                .orElseThrow(() -> new ThemeNotFoundException(adminReservationRequest.themeId()));
+
+        Member member = memberRepository.findById(adminReservationRequest.memberId())
+                .orElseThrow(() -> new MemberNotFoundException(adminReservationRequest.memberId()));
+
+        Reservation newReservation = new Reservation(adminReservationRequest.date(),
+                reservationTime, theme, member);
+
+        return ReservationResponse.from(reservationRepository.save(newReservation));
+    }
+
+    public List<ReservationResponse> getReservationsByConditions(
+            @Valid SearchConditionsRequest searchConditionsRequest) {
+
+        List<Reservation> reservations = reservationRepository.findAllByThemeIdAndMemberIdAndDateBetween(
+                searchConditionsRequest.themeId(),
+                searchConditionsRequest.memberId(),
+                searchConditionsRequest.dateFrom(),
+                searchConditionsRequest.dateTo()
+        );
+        return reservations.stream().
+                map(reservation -> ReservationResponse.from(reservation))
+                .toList();
+    }
+
+    public List<MemberReservationResponse> getReservationByMember(Member member) {
+        return reservationRepository.findAllByMember(member)
+                .stream()
+                .map(MemberReservationResponse::from)
+                .toList();
+    }
 }
