+++ src/test/java/roomescape/service/theme/ThemeServiceTest.java
@@ -1,23 +1,17 @@
 package roomescape.service.theme;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.when;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.exception.theme.ThemeNotFoundException;
 import roomescape.repository.theme.ThemeRepository;
 
 @ExtendWith(MockitoExtension.class)
-class ThemeServiceImplTest {
+class ThemeServiceTest {
     @Mock
     private ThemeRepository themeRepository;
 
     @InjectMocks
-    private ThemeServiceImpl themeService;
+    private ThemeService themeService;
 
 }
